requirements.txt
# Core
python-dotenv
requests
pytest
Flask

# Other
pandas
numpy
selenium
pytesseract
openai
webdriver-manager